Restructure main and first command
